Move storage package under registry package
