+++ registry/handlers/app.go
@@ -11,8 +11,8 @@ import (
 	"github.com/docker/distribution/registry/auth"
 	"github.com/docker/distribution/configuration"
 	ctxu "github.com/docker/distribution/context"
-	"github.com/docker/distribution/storage"
-	"github.com/docker/distribution/storage/notifications"
+	"github.com/docker/distribution/registry/storage"
+	"github.com/docker/distribution/registry/storage/notifications"
 	"github.com/docker/distribution/storagedriver"
 	"github.com/docker/distribution/storagedriver/factory"
 	"github.com/gorilla/mux"
