+++ registry/handlers/app_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/docker/distribution/registry/api/v2"
 	_ "github.com/docker/distribution/registry/auth/silly"
 	"github.com/docker/distribution/configuration"
-	"github.com/docker/distribution/storage"
+	"github.com/docker/distribution/registry/storage"
 	"github.com/docker/distribution/storagedriver/inmemory"
 	"golang.org/x/net/context"
 )
