+++ registry/handlers/context.go
@@ -7,7 +7,7 @@ import (
 	"github.com/docker/distribution/registry/api/v2"
 	ctxu "github.com/docker/distribution/context"
 	"github.com/docker/distribution/digest"
-	"github.com/docker/distribution/storage"
+	"github.com/docker/distribution/registry/storage"
 	"golang.org/x/net/context"
 )
