+++ registry/storage/cloudfrontlayerhandler.go
