+++ registry/storage/doc.go
