+++ registry/storage/filereader.go
