+++ registry/storage/filereader_test.go
