+++ registry/storage/filewriter.go
