+++ registry/storage/filewriter_test.go
