+++ registry/storage/layerreader.go
