+++ registry/storage/layerstore.go
