+++ registry/storage/manifeststore_test.go
