+++ registry/storage/paths.go
