+++ registry/storage/paths_test.go
