+++ registry/storage/registry.go
